Major new updates:
The construct relation has been removed from the analysis. Explorative examination has found that the construct relation introduces too much noise into the results. This was noticed because an earlier iteration of this dataset had extra requirements for headship on construct relations that restricted the number of constructs and kept them low. This restriction was an accident caused by a poorly-worded search template. When the templates were edited, many more construct relations entered into the dataset. The outlines of the data disappeared, and things became chaotic. Futher analysis of the construct relations revealed a lot of spurious connections between words made on the basis of construct relations. Because constructs are more semantically flexible, they seem to be unreliable at the moment for gathering good semantic data.

Other updates include new gephi graphs, new PCA plots, and a new spreadsheets directory (which includes .csv data for the paper).